Move utils from grafana-ui to grafana-data
