+++ packages/grafana-data/src/utils/flotPairs.ts
@@ -1,5 +1,10 @@
+import { Field } from '../types/dataFrame';
+import { NullValueMode } from '../types/data';
+import { GraphSeriesValue } from '../types/graph';
+import { TimeRange } from '../types/time';
+
 // Types
-import { NullValueMode, GraphSeriesValue, Field, TimeRange } from '@grafana/data';
+// import { NullValueMode, GraphSeriesValue, Field, TimeRange } from '@grafana/data';
 
 export interface FlotPairsOptions {
   xField: Field;
