+++ packages/grafana-ui/src/utils/index.ts
@@ -1,9 +1,5 @@
 export * from './colors';
-export * from './namedColorsPalette';
-export * from './displayProcessor';
-export * from './fieldDisplay';
 export * from './validate';
-export { getFlotPairs, getFlotPairsConstant } from './flotPairs';
 export * from './slate';
 export * from './dataLinks';
 export * from './tags';
